+++ src/pages/offer-page/offer-page.tsx
@@ -1,8 +1,6 @@
 import { Helmet } from 'react-helmet-async';
 import { DetailedOffers } from '../../types/offers';
 import { useParams } from 'react-router-dom';
-import Header from '../../components/header/header';
-import { AuthorizationStatus } from '../../consts';
 import { calculateRatingWidth, capitalizeWord } from '../../utils';
 import ReviewForm from '../../components/review-form/review-form';
 import Page404 from '../page-404/page-404';
@@ -21,244 +19,240 @@ function OfferPage({offers}: OfferProps): JSX.Element {
   }
 
   return (
-    <div className="page">
+    <main className="page__main page__main--offer">
       <Helmet>
         <title>6 cities: Offer</title>
       </Helmet>
-      <Header offers={offers} authorizationStatus={AuthorizationStatus.NoAuth} />
-
-      <main className="page__main page__main--offer">
-        <section className="offer">
-          <div className="offer__gallery-container container">
-            <div className="offer__gallery">
-              <div className="offer__image-wrapper">
-                <img className="offer__image" src="img/room.jpg" alt="Photo studio" />
-              </div>
-              <div className="offer__image-wrapper">
-                <img className="offer__image" src="img/apartment-01.jpg" alt="Photo studio" />
-              </div>
-              <div className="offer__image-wrapper">
-                <img className="offer__image" src="img/apartment-02.jpg" alt="Photo studio" />
-              </div>
-              <div className="offer__image-wrapper">
-                <img className="offer__image" src="img/apartment-03.jpg" alt="Photo studio" />
-              </div>
-              <div className="offer__image-wrapper">
-                <img className="offer__image" src="img/studio-01.jpg" alt="Photo studio" />
-              </div>
-              <div className="offer__image-wrapper">
-                <img className="offer__image" src="img/apartment-01.jpg" alt="Photo studio" />
-              </div>
+      <section className="offer">
+        <div className="offer__gallery-container container">
+          <div className="offer__gallery">
+            <div className="offer__image-wrapper">
+              <img className="offer__image" src="img/room.jpg" alt="Photo studio" />
+            </div>
+            <div className="offer__image-wrapper">
+              <img className="offer__image" src="img/apartment-01.jpg" alt="Photo studio" />
+            </div>
+            <div className="offer__image-wrapper">
+              <img className="offer__image" src="img/apartment-02.jpg" alt="Photo studio" />
+            </div>
+            <div className="offer__image-wrapper">
+              <img className="offer__image" src="img/apartment-03.jpg" alt="Photo studio" />
+            </div>
+            <div className="offer__image-wrapper">
+              <img className="offer__image" src="img/studio-01.jpg" alt="Photo studio" />
+            </div>
+            <div className="offer__image-wrapper">
+              <img className="offer__image" src="img/apartment-01.jpg" alt="Photo studio" />
             </div>
           </div>
-          <div className="offer__container container">
-            <div className="offer__wrapper">
-              {currentOffer.isPremium &&
+        </div>
+        <div className="offer__container container">
+          <div className="offer__wrapper">
+            {currentOffer.isPremium &&
               <div className="offer__mark">
                 <span>Premium</span>
               </div>}
-              <div className="offer__name-wrapper">
-                <h1 className="offer__name">
-                  {currentOffer.title}
-                </h1>
-                <button className={`offer__bookmark-button ${isFavorite} button`} type="button">
-                  <svg className="offer__bookmark-icon" width="31" height="33">
-                    <use xlinkHref="#icon-bookmark"></use>
-                  </svg>
-                  <span className="visually-hidden">To bookmarks</span>
-                </button>
+            <div className="offer__name-wrapper">
+              <h1 className="offer__name">
+                {currentOffer.title}
+              </h1>
+              <button className={`offer__bookmark-button ${isFavorite} button`} type="button">
+                <svg className="offer__bookmark-icon" width="31" height="33">
+                  <use xlinkHref="#icon-bookmark"></use>
+                </svg>
+                <span className="visually-hidden">To bookmarks</span>
+              </button>
+            </div>
+            <div className="offer__rating rating">
+              <div className="offer__stars rating__stars">
+                <span style={{width: calculateRatingWidth(currentOffer.rating)}}></span>
+                <span className="visually-hidden">Rating</span>
               </div>
-              <div className="offer__rating rating">
-                <div className="offer__stars rating__stars">
-                  <span style={{width: calculateRatingWidth(currentOffer.rating)}}></span>
-                  <span className="visually-hidden">Rating</span>
+              <span className="offer__rating-value rating__value">{currentOffer.rating}</span>
+            </div>
+            <ul className="offer__features">
+              <li className="offer__feature offer__feature--entire">
+                {capitalizeWord(currentOffer.type)}
+              </li>
+              <li className="offer__feature offer__feature--bedrooms">
+                {(currentOffer.bedrooms > 1) ? `${currentOffer.bedrooms} Bedrooms` : `${currentOffer.bedrooms} Bedroom`}
+              </li>
+              <li className="offer__feature offer__feature--adults">
+                {(currentOffer.maxAdults > 1) ? `Max ${currentOffer.maxAdults} Adults` : `Max ${currentOffer.maxAdults} Adult`}
+              </li>
+            </ul>
+            <div className="offer__price">
+              <b className="offer__price-value">&euro;{currentOffer.price}</b>
+              <span className="offer__price-text">&nbsp;night</span>
+            </div>
+            <div className="offer__inside">
+              <h2 className="offer__inside-title">What&apos;s inside</h2>
+              <ul className="offer__inside-list">
+                {currentOffer.goods.map((feature) => (
+                  <li className="offer__inside-item" key={feature + currentOffer.id}>
+                    {feature}
+                  </li>)
+                )}
+              </ul>
+            </div>
+            <div className="offer__host">
+              <h2 className="offer__host-title">Meet the host</h2>
+              <div className="offer__host-user user">
+                <div className="offer__avatar-wrapper offer__avatar-wrapper--pro user__avatar-wrapper">
+                  <img className="offer__avatar user__avatar" src="img/avatar-angelina.jpg" width="74" height="74" alt="Host avatar" />
                 </div>
-                <span className="offer__rating-value rating__value">{currentOffer.rating}</span>
+                <span className="offer__user-name">
+                  {currentOffer.host.name}
+                </span>
+                {currentOffer.host.isPro && <span className="offer__user-status">Pro</span>}
               </div>
-              <ul className="offer__features">
-                <li className="offer__feature offer__feature--entire">
-                  {capitalizeWord(currentOffer.type)}
-                </li>
-                <li className="offer__feature offer__feature--bedrooms">
-                  {(currentOffer.bedrooms > 1) ? `${currentOffer.bedrooms} Bedrooms` : `${currentOffer.bedrooms} Bedroom`}
-                </li>
-                <li className="offer__feature offer__feature--adults">
-                  {(currentOffer.maxAdults > 1) ? `Max ${currentOffer.maxAdults} Adults` : `Max ${currentOffer.maxAdults} Adult`}
+              <div className="offer__description">
+                <p className="offer__text">
+                  {currentOffer.description}
+                </p>
+                <p className="offer__text">
+                    An independent House, strategically located between Rembrand Square and National Opera, but where the bustle of the city comes to rest in this alley flowery and colorful.
+                </p>
+              </div>
+            </div>
+            <section className="offer__reviews reviews">
+              <h2 className="reviews__title">Reviews &middot; <span className="reviews__amount">1</span></h2>
+              <ul className="reviews__list">
+                <li className="reviews__item">
+                  <div className="reviews__user user">
+                    <div className="reviews__avatar-wrapper user__avatar-wrapper">
+                      <img className="reviews__avatar user__avatar" src="img/avatar-max.jpg" width="54" height="54" alt="Reviews avatar" />
+                    </div>
+                    <span className="reviews__user-name">
+                        Max
+                    </span>
+                  </div>
+                  <div className="reviews__info">
+                    <div className="reviews__rating rating">
+                      <div className="reviews__stars rating__stars">
+                        <span style={{width: '80%'}}></span>
+                        <span className="visually-hidden">Rating</span>
+                      </div>
+                    </div>
+                    <p className="reviews__text">
+                        A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam. The building is green and from 18th century.
+                    </p>
+                    <time className="reviews__time" dateTime="2019-04-24">April 2019</time>
+                  </div>
                 </li>
               </ul>
-              <div className="offer__price">
-                <b className="offer__price-value">&euro;{currentOffer.price}</b>
-                <span className="offer__price-text">&nbsp;night</span>
-              </div>
-              <div className="offer__inside">
-                <h2 className="offer__inside-title">What&apos;s inside</h2>
-                <ul className="offer__inside-list">
-                  {currentOffer.goods.map((feature) => (
-                    <li className="offer__inside-item" key={feature + currentOffer.id}>
-                      {feature}
-                    </li>)
-                  )}
-                </ul>
-              </div>
-              <div className="offer__host">
-                <h2 className="offer__host-title">Meet the host</h2>
-                <div className="offer__host-user user">
-                  <div className="offer__avatar-wrapper offer__avatar-wrapper--pro user__avatar-wrapper">
-                    <img className="offer__avatar user__avatar" src="img/avatar-angelina.jpg" width="74" height="74" alt="Host avatar" />
-                  </div>
-                  <span className="offer__user-name">
-                    {currentOffer.host.name}
-                  </span>
-                  {currentOffer.host.isPro && <span className="offer__user-status">Pro</span>}
-                </div>
-                <div className="offer__description">
-                  <p className="offer__text">
-                    {currentOffer.description}
-                  </p>
-                  <p className="offer__text">
-                    An independent House, strategically located between Rembrand Square and National Opera, but where the bustle of the city comes to rest in this alley flowery and colorful.
-                  </p>
-                </div>
-              </div>
-              <section className="offer__reviews reviews">
-                <h2 className="reviews__title">Reviews &middot; <span className="reviews__amount">1</span></h2>
-                <ul className="reviews__list">
-                  <li className="reviews__item">
-                    <div className="reviews__user user">
-                      <div className="reviews__avatar-wrapper user__avatar-wrapper">
-                        <img className="reviews__avatar user__avatar" src="img/avatar-max.jpg" width="54" height="54" alt="Reviews avatar" />
-                      </div>
-                      <span className="reviews__user-name">
-                        Max
-                      </span>
-                    </div>
-                    <div className="reviews__info">
-                      <div className="reviews__rating rating">
-                        <div className="reviews__stars rating__stars">
-                          <span style={{width: '80%'}}></span>
-                          <span className="visually-hidden">Rating</span>
-                        </div>
-                      </div>
-                      <p className="reviews__text">
-                        A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam. The building is green and from 18th century.
-                      </p>
-                      <time className="reviews__time" dateTime="2019-04-24">April 2019</time>
-                    </div>
-                  </li>
-                </ul>
-                <ReviewForm />
-              </section>
-            </div>
+              <ReviewForm />
+            </section>
           </div>
-          <section className="offer__map map"></section>
-        </section>
-        <div className="container">
-          <section className="near-places places">
-            <h2 className="near-places__title">Other places in the neighbourhood</h2>
-            <div className="near-places__list places__list">
-              <article className="near-places__card place-card">
-                <div className="near-places__image-wrapper place-card__image-wrapper">
-                  <a href="#">
-                    <img className="place-card__image" src="img/room.jpg" width="260" height="200" alt="Place image" />
-                  </a>
-                </div>
-                <div className="place-card__info">
-                  <div className="place-card__price-wrapper">
-                    <div className="place-card__price">
-                      <b className="place-card__price-value">&euro;80</b>
-                      <span className="place-card__price-text">&#47;&nbsp;night</span>
-                    </div>
-                    <button className="place-card__bookmark-button place-card__bookmark-button--active button" type="button">
-                      <svg className="place-card__bookmark-icon" width="18" height="19">
-                        <use xlinkHref="#icon-bookmark"></use>
-                      </svg>
-                      <span className="visually-hidden">In bookmarks</span>
-                    </button>
-                  </div>
-                  <div className="place-card__rating rating">
-                    <div className="place-card__stars rating__stars">
-                      <span style={{width: '80%'}}></span>
-                      <span className="visually-hidden">Rating</span>
-                    </div>
-                  </div>
-                  <h2 className="place-card__name">
-                    <a href="#">Wood and stone place</a>
-                  </h2>
-                  <p className="place-card__type">Room</p>
-                </div>
-              </article>
-
-              <article className="near-places__card place-card">
-                <div className="near-places__image-wrapper place-card__image-wrapper">
-                  <a href="#">
-                    <img className="place-card__image" src="img/apartment-02.jpg" width="260" height="200" alt="Place image" />
-                  </a>
-                </div>
-                <div className="place-card__info">
-                  <div className="place-card__price-wrapper">
-                    <div className="place-card__price">
-                      <b className="place-card__price-value">&euro;132</b>
-                      <span className="place-card__price-text">&#47;&nbsp;night</span>
-                    </div>
-                    <button className="place-card__bookmark-button button" type="button">
-                      <svg className="place-card__bookmark-icon" width="18" height="19">
-                        <use xlinkHref="#icon-bookmark"></use>
-                      </svg>
-                      <span className="visually-hidden">To bookmarks</span>
-                    </button>
-                  </div>
-                  <div className="place-card__rating rating">
-                    <div className="place-card__stars rating__stars">
-                      <span style={{width: '80%'}}></span>
-                      <span className="visually-hidden">Rating</span>
-                    </div>
-                  </div>
-                  <h2 className="place-card__name">
-                    <a href="#">Canal View Prinsengracht</a>
-                  </h2>
-                  <p className="place-card__type">Apartment</p>
-                </div>
-              </article>
-
-              <article className="near-places__card place-card">
-                <div className="place-card__mark">
-                  <span>Premium</span>
-                </div>
-                <div className="near-places__image-wrapper place-card__image-wrapper">
-                  <a href="#">
-                    <img className="place-card__image" src="img/apartment-03.jpg" width="260" height="200" alt="Place image" />
-                  </a>
-                </div>
-                <div className="place-card__info">
-                  <div className="place-card__price-wrapper">
-                    <div className="place-card__price">
-                      <b className="place-card__price-value">&euro;180</b>
-                      <span className="place-card__price-text">&#47;&nbsp;night</span>
-                    </div>
-                    <button className="place-card__bookmark-button button" type="button">
-                      <svg className="place-card__bookmark-icon" width="18" height="19">
-                        <use xlinkHref="#icon-bookmark"></use>
-                      </svg>
-                      <span className="visually-hidden">To bookmarks</span>
-                    </button>
-                  </div>
-                  <div className="place-card__rating rating">
-                    <div className="place-card__stars rating__stars">
-                      <span style={{width: '100%'}}></span>
-                      <span className="visually-hidden">Rating</span>
-                    </div>
-                  </div>
-                  <h2 className="place-card__name">
-                    <a href="#">Nice, cozy, warm big bed apartment</a>
-                  </h2>
-                  <p className="place-card__type">Apartment</p>
-                </div>
-              </article>
-            </div>
-          </section>
         </div>
-      </main>
-    </div>
+        <section className="offer__map map"></section>
+      </section>
+      <div className="container">
+        <section className="near-places places">
+          <h2 className="near-places__title">Other places in the neighbourhood</h2>
+          <div className="near-places__list places__list">
+            <article className="near-places__card place-card">
+              <div className="near-places__image-wrapper place-card__image-wrapper">
+                <a href="#">
+                  <img className="place-card__image" src="img/room.jpg" width="260" height="200" alt="Place image" />
+                </a>
+              </div>
+              <div className="place-card__info">
+                <div className="place-card__price-wrapper">
+                  <div className="place-card__price">
+                    <b className="place-card__price-value">&euro;80</b>
+                    <span className="place-card__price-text">&#47;&nbsp;night</span>
+                  </div>
+                  <button className="place-card__bookmark-button place-card__bookmark-button--active button" type="button">
+                    <svg className="place-card__bookmark-icon" width="18" height="19">
+                      <use xlinkHref="#icon-bookmark"></use>
+                    </svg>
+                    <span className="visually-hidden">In bookmarks</span>
+                  </button>
+                </div>
+                <div className="place-card__rating rating">
+                  <div className="place-card__stars rating__stars">
+                    <span style={{width: '80%'}}></span>
+                    <span className="visually-hidden">Rating</span>
+                  </div>
+                </div>
+                <h2 className="place-card__name">
+                  <a href="#">Wood and stone place</a>
+                </h2>
+                <p className="place-card__type">Room</p>
+              </div>
+            </article>
+
+            <article className="near-places__card place-card">
+              <div className="near-places__image-wrapper place-card__image-wrapper">
+                <a href="#">
+                  <img className="place-card__image" src="img/apartment-02.jpg" width="260" height="200" alt="Place image" />
+                </a>
+              </div>
+              <div className="place-card__info">
+                <div className="place-card__price-wrapper">
+                  <div className="place-card__price">
+                    <b className="place-card__price-value">&euro;132</b>
+                    <span className="place-card__price-text">&#47;&nbsp;night</span>
+                  </div>
+                  <button className="place-card__bookmark-button button" type="button">
+                    <svg className="place-card__bookmark-icon" width="18" height="19">
+                      <use xlinkHref="#icon-bookmark"></use>
+                    </svg>
+                    <span className="visually-hidden">To bookmarks</span>
+                  </button>
+                </div>
+                <div className="place-card__rating rating">
+                  <div className="place-card__stars rating__stars">
+                    <span style={{width: '80%'}}></span>
+                    <span className="visually-hidden">Rating</span>
+                  </div>
+                </div>
+                <h2 className="place-card__name">
+                  <a href="#">Canal View Prinsengracht</a>
+                </h2>
+                <p className="place-card__type">Apartment</p>
+              </div>
+            </article>
+
+            <article className="near-places__card place-card">
+              <div className="place-card__mark">
+                <span>Premium</span>
+              </div>
+              <div className="near-places__image-wrapper place-card__image-wrapper">
+                <a href="#">
+                  <img className="place-card__image" src="img/apartment-03.jpg" width="260" height="200" alt="Place image" />
+                </a>
+              </div>
+              <div className="place-card__info">
+                <div className="place-card__price-wrapper">
+                  <div className="place-card__price">
+                    <b className="place-card__price-value">&euro;180</b>
+                    <span className="place-card__price-text">&#47;&nbsp;night</span>
+                  </div>
+                  <button className="place-card__bookmark-button button" type="button">
+                    <svg className="place-card__bookmark-icon" width="18" height="19">
+                      <use xlinkHref="#icon-bookmark"></use>
+                    </svg>
+                    <span className="visually-hidden">To bookmarks</span>
+                  </button>
+                </div>
+                <div className="place-card__rating rating">
+                  <div className="place-card__stars rating__stars">
+                    <span style={{width: '100%'}}></span>
+                    <span className="visually-hidden">Rating</span>
+                  </div>
+                </div>
+                <h2 className="place-card__name">
+                  <a href="#">Nice, cozy, warm big bed apartment</a>
+                </h2>
+                <p className="place-card__type">Apartment</p>
+              </div>
+            </article>
+          </div>
+        </section>
+      </div>
+    </main>
   );
 }
